refactor: consolidate DragMover into Mover
